requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-google-genai>=0.0.8
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0
pypdf>=4.0.0
python-dotenv>=1.0.0
langchain-community>=0.0.10
